Finished Health Potion
The health potion is complete. The player can equip the potion by simply walking into it. The potion can be used if the player holds "E". The potion restores 25 health to the user and is removed once used. If the user attempts to use the potion while their health is full, it will not work.